Require the user to specify cxfVersion and cxfPluginVersion
The current default value allows the plugin to grab an arbitrary version
of CXF. It is a time bomb for people who don't read the docs too
closely, have missed what the default is and didn't override it with a
specific version. Ideally, people should read the documentation, but the
reality is a touch different.

It seems that gradle automatically treats `@Input`s as non-nullable and
complains if any of them don't have a value, which is nice.

This is a breaking change for those who haven't specified these
parameters before. I don't think the breakage will be much worse than
what the release of CXF 4.0.0 at 22nd of December has done.

Alternatives: emphasize the danger of the default value in the README,
pick some fixed version arbitrarily. The former would help but still
leaves the human factor very relevant. The latter feels wrong, it's
probably not the choice the plugin should make. It also can lead to the
same kind of breakage as a CXF release now.